# core-banking-payments-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-baking-payments
  description: Payments Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/payments
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/payment-orders/{paymentOrderId}/proof/{paymentProofId}:
    get:
      tags:
        - Payment Proofs
      summary: Get Payment Proof by ID
      description: Retrieve a specific payment proof by its unique identifier, within the context of a payment order.
      operationId: getPaymentProofById
      parameters:
        - name: paymentOrderId
          in: path
          description: Unique identifier of the payment order
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentProofId
          in: path
          description: Unique identifier of the payment proof
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the payment proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentProofDTO'
        '404':
          description: Payment proof not found
    put:
      tags:
        - Payment Proofs
      summary: Update Payment Proof
      description: Update an existing payment proof by its unique identifier.
      operationId: updatePaymentProof
      parameters:
        - name: paymentOrderId
          in: path
          description: Unique identifier of the payment order
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentProofId
          in: path
          description: Unique identifier of the payment proof to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentProofDTO'
      responses:
        '200':
          description: Payment proof updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentProofDTO'
        '404':
          description: Payment proof not found
    delete:
      tags:
        - Payment Proofs
      summary: Delete Payment Proof
      description: Remove an existing payment proof by its unique identifier, within the context of a payment order.
      operationId: deletePaymentProof
      parameters:
        - name: paymentOrderId
          in: path
          description: Unique identifier of the payment order
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentProofId
          in: path
          description: Unique identifier of the payment proof to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Payment proof deleted successfully
        '404':
          description: Payment proof not found
...
  /api/v1/payment-orders/{paymentOrderId}/proof:
    get:
      tags:
        - Payment Proofs
      summary: Retrieve Payment Proofs
      description: Retrieves a paginated list of payment proofs for a specific payment order, with optional filtering.
      operationId: getAllPaymentProofs
      parameters:
        - name: paymentOrderId
          in: path
          description: Unique identifier of the payment order
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.paymentOrderId
          in: query
          description: Exact filter for payment order id
          required: false
          schema:
            type: string
        - name: filters.documentId
          in: query
          description: Exact filter for document id
          required: false
          schema:
            type: string
        - name: filters.proofType
          in: query
          description: Exact filter for proof type
          required: false
          schema:
            type: string
        - name: filters.proofDate
          in: query
          description: Exact filter for proof date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[proofDate].from
          in: query
          description: Filter proof date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[proofDate].to
          in: query
          description: Filter proof date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateCreated
          in: query
          description: Exact filter for date created
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].from
          in: query
          description: Filter date created from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].to
          in: query
          description: Filter date created to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateUpdated
          in: query
          description: Exact filter for date updated
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].from
          in: query
          description: Filter date updated from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].to
          in: query
          description: Filter date updated to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved payment proofs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: No payment proofs found
    post:
      tags:
        - Payment Proofs
      summary: Create Payment Proof
      description: Creates a new payment proof for the specified payment order.
      operationId: createPaymentProof
      parameters:
        - name: paymentOrderId
          in: path
          description: Unique identifier of the payment order
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentProofDTO'
      responses:
        '201':
          description: Payment proof created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentProofDTO'
        '400':
          description: Invalid payment proof data provided
...
components:
  schemas:
...
    PaymentProofDTO:
      required:
        - documentId
        - paymentOrderId
        - proofDate
        - proofType
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        paymentProofId:
          type: string
          format: uuid
          readOnly: true
        paymentOrderId:
          type: string
          format: uuid
        documentId:
          type: string
          format: uuid
        proofType:
          maxLength: 50
          minLength: 0
          type: string
        proofDate:
          type: string
          format: date-time
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...